function-app/requirements.txt
pandas
numpy
azure-functions
azure-storage-blob
openpyxl
